refactor: merge everything back into __main__
